+++ src/Dap.Gui/Internal/GuiApp.fs
@@ -9,8 +9,11 @@ open Dap.Platform
 open Dap.Gui
 
 type GuiApp (app : IBaseApp) =
-    static let mutable instance : IGuiApp option = None
+    static let mutable instance : GuiApp option = None
     let logger = app.Env.Logging.GetLogger "GuiApp"
+    let mutable state : GuiAppState = Foreground
+    let onWillChangeState = new Bus<GuiAppState> (app.Env, "OnWillChangeState")
+    let onDidChangeState = new Bus<unit> (app.Env, "OnDidChangeState")
     let mutable hooks : IGuiAppHook list = []
     let mutable themes : Map<string, ITheme> = Map.empty
     let mutable theme : ITheme option = None
@@ -20,7 +23,7 @@ type GuiApp (app : IBaseApp) =
         if theme.IsNone || theme.Value.Key <> theme'.Key then
             if theme.IsSome then
                 onWillSwitchTheme.Trigger theme'
-                logWarn logger "SwitchTheme" theme'.Key (theme, theme')
+                logWarn logger "SetTheme" theme'.Key (theme, theme')
             else
                 logWarn logger "SetTheme" theme'.Key (theme')
             theme <- Some theme'
@@ -33,10 +36,18 @@ type GuiApp (app : IBaseApp) =
             logError logger "Singleton_Violated" "Instance_Exist" (instance.Value, this)
         else
             let instance' = this :> IGuiApp
-            instance <- Some instance'
-            hooks <- Hook.createAll<IGuiAppHook> instance'.App.Env.Logging
+            instance <- Some this
+            hooks <- Hook.createAll<IGuiAppHook> app.Env.Logging
             hooks
             |> List.iter (fun x -> x.OnInit instance')
+    member __.SetState' (state' : GuiAppState) =
+        if state <> state' then
+            onWillChangeState.Trigger state'
+            logWarn logger "SetState" (state'.ToString ()) (state)
+            state <- state'
+            onDidChangeState.Trigger ()
+    member __.App = app
+    member __.State = state
     member __.Theme =
         if theme.IsNone then
             setTheme (new Theme (NoKey, null) :> ITheme)
@@ -56,6 +67,9 @@ type GuiApp (app : IBaseApp) =
             )
     interface IGuiApp with
         member __.App = app
+        member this.State = this.State
+        member __.OnWillChangeState = onWillChangeState.Publish
+        member __.OnDidChangeState = onDidChangeState.Publish
         member this.Theme = this.Theme
         member __.AddTheme (key : string) (param : 'param) (setup : ITheme -> 'param -> unit) =
             themes